+++ views/widget_view.h
@@ -1,8 +1,14 @@
 #ifndef WIDGET_VIEW
 #define WIDGET_VIEW
 
+/* ------------------------------------------------------------------------------------------------------------------ */
+/* Public Function Declarations                                                                                       */
+/* ------------------------------------------------------------------------------------------------------------------ */
+
 void widget_view_alloc(void);
 void widget_view_free(void);
 void widget_view_switch(void);
+void widget_reset_switch(void);
+void widget_about_switch(void);
 
 #endif