+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Dolphin State changelog
-## Dolphin State 1.0
+## Dolphin State v1.1
+- Added main menu
+- Added About information
+- Added possibility to reset daily limits
+## Dolphin State v1.0
 - Initial release